#32: Changed the parsing architecture to return only a stubbed structure, not actual MUIS elements.  A separate "content creator" creates the actual elements and initializes them now.
It seems that this change has triggered a bug in the style sheet code that is causing the animated demo to make a green button on a green background.  The animated style expression maybe isn't getting enough priority.  I don't think I actually created the bug with these changes--just exposed it.