#32: More work on templates:
I refactored the role attribute into the TemplateStructure class, so now there is a role attribute per template type.  As noted in the comment on the issue, the attribute manager does not currently support multiple attributes with the same name.  I will need to correct this before any of this works.

Implemented the AttachPointInstance inner class which enables basic get/set/collection functionality for attach points.  Only one particularly difficult piece of logic missing on the setter for the element occupying a singular attach point.  I left a compile error to mark it.

Not confident in how I tied up the initChildren functionality.  I need to try to follow that logic through before I start debugging.