#7: Utilized the simple cache to improve performance on FlowLayout. Helped a lot.  The layout is at least as fast as the render now.
It's not perfect.  At a particular width a part of the first content is
hidden.  Also I haven't tested it with images, or more than 2 elements.
It needs a larger test case.

I can't decide whether to hammer it out well or just call it good enough
for now and move on to some more fun stuff.